Add nval and scoreval pruning to spectral clustering
Previously a fixed percentage of the total sample kept for the affinity
scoring, with the rest set to 0. This leads to varying amount of samples
kept depending on audio length. This results in dropping of speakers
with short duration in a long audio file.

nval keeps a fixed number of sample, making it robust to changing
audio duration and speaker count.